--- a/pastas/decorators.py
+++ b/pastas/decorators.py
@@ -146,7 +146,7 @@ def deprecate_args_or_kwargs(
         raise DeprecationWarning(msg)
 
 
-def njit(function: Optional[Function] = None, parallel: bool = False) -> Function:
+def njit(function: Optional[Function] = None, **kwargs) -> Function:
     def njit_decorator(f: Function):
         try:
             if not USE_NUMBA:
@@ -154,7 +154,7 @@ def njit(function: Optional[Function] = None, parallel: bool = False) -> Functio
             else:
                 from numba import njit
 
-                fnumba = njit(f, parallel=parallel)
+                fnumba = njit(f, **kwargs)
                 return fnumba
         except ImportError:
             return f