This PR introduces some breaking changes, and I'm not sure how to deal with that:
 - Previous code `ps.stats.ccf(a, b)` is now calculated as `ps.stats.ccf(b, a)`. Do we add a warning?
 - The 0 lag for the ccf is returned. I think it's okay to just return the new result. 

The ACF is unaffected. One slight issue is that I'm getting ACFs > 1.0 warnings in the autocorrelation notebook. The max exceedance is <1e-3 so it's a computational issue having to do with the binning methods, as the current warning suggests. Not sure what to do about that...  

I don't think many people will be affected by these breaking changes since I don't think many people were actively using the CCF except for computing the autocorrelation, which is unchanged. So I think it would be okay to include these changes, and perhaps include a warning for flipped order of x and y.

# Short description
Deals with the following issues:
- #846 
- #912 
- #915 
- #917

### Lag 0 for CCF
Lag 0 for the CCF is now returned. In `ps.stats.acf()` this value is dropped if lags is passed as an integer. 

### Flip order of x, y to match numpy/scipy.correlate

The results now match numpy/scipy: `ps.stats.ccf(a, b)` is similar to `np.correlate(a, b, mode="full")[len(a) - 1:]`

### Performance
Improves performance considerably:
- Traverse only the lower diagonal of the dot product between x and y, in other words only consider the lags in one direction.
- In `_compute_ccf_gaussian`, check whether abs(time lag) is smaller than 6 *  std.dev. Beyond that limit the computed weight becomes negligible. See comment below.
- Use `numba.prange` for the outer loop
- Set `parallel=True, nogil=True, cache=True` in `@njit` decorator.
- Minor performance improvements (e.g. array access once per loop, etc.).

Many thanks @AviadAvr for suggesting the implementation of `numba.prange`!

Timings (N=1000, lags=365):
For `_compute_ccf_regular()`
```
Original       : 26.2 ms ± 1.6 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
This PR        : 4.25 ms ± 5.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <-- I reverted this change
Pastas plugins : 993 µs ± 133 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
EDIT: unfortunately, `njit`-ing the `_compute_ccf_regular` made it rather flaky. I was getting significant speedups, but the `test_acf_hourly` went from very fast to ~15s. So I reverted that change. For very fast CCFS for time series with regular timesteps `pastas_plugins` is recommended. 


For `_compute_ccf_gaussian()`
```
Original       : 4.51 s ± 69.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
This PR        : 559 ms ± 16.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

For `_compute_ccf_rectangle()`
```
Original       : 248 ms ± 1.85 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
This PR        : 56.1 ms ± 1.86 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

# Checklist before PR can be merged:
- [x] closes issues
   - [x] #846 
   - [x] #912 
   - [x] #915 
   - [x] #917
- [x] is documented
- [x] code formatted and linted with [ruff](https://docs.astral.sh/ruff/)
- [x] type hints
- [x] tests added or expanded