@raoulcollenteur 2025-03-05
Thanks for this PR, looks nice! I think we can safely reverse the order for the CCF to match scipy/numpy, I don't think many/any people are using this function. And if so, they probably know what they are doing and will note changes that can be tracked back to this PR.

Regarding the use of `numba.prange` (here, but also elsewhere), I wanted to implement this a while back but @martinvonk  mentioned that this could cause problems when running multiple pastas models in parallel (i.e., with pastastore). Does this mean the use of `prange` does not cause any issues and we can implement it in other parts of Pastas too? 

Perhaps we should only raise the warning when acf > |1.01| or somewhere around that? 

Good to go I think if prange is safe to use now :)