@dbrakenhoff 2025-03-06
> Regarding the use of numba.prange (here, but also elsewhere), I wanted to implement this a while back but @martinvonk mentioned that this could cause problems when running multiple pastas models in parallel (i.e., with pastastore). Does this mean the use of prange does not cause any issues and we can implement it in other parts of Pastas too?

Pastastore doesnt´t currently provide any direct methods for computing the ACF/CCF in parallel. So implementing it here isn't much of a risk. Additionally, I just found out that `numba.set_num_threads(1)` can be used to make parallel jitted functions run in a single thread. We might have to explicitly start using that in pastastore if it turns out that numba parallel affects the naive parallelization implemented in pastastore, but I think it's okay to start implementing it. 

> Perhaps we should only raise the warning when acf > |1.01| or somewhere around that?

Yeah, agreed, I'll modify the threshold.