@martinvonk 2025-03-06
I think, before making stuff parallel in Pastas we should first make a discussion / issue on how and what. Think this through a bit. 

Because maybe we want PastaStore non-parallel and Pastas in parallel or the other way round?
Maybe the default for Pastas should be to run everything not in parallel? Add something to the rcParams["num_cores"] = 1?
Do we need if `__name__ == "__main__"` statements to ensure memory safety? 
How about linux/macos/windows differences for parallel?