@dbrakenhoff 2025-03-06
> Because maybe we want PastaStore non-parallel and Pastas in parallel or the other way round?

It'll probably end up as a bit of both. As long as it's easy to configure, i.e. turn parallel off in both modules, it should be simple to accommodate both approaches. And so far, that seems simple to do. 

> Maybe the default for Pastas should be to run everything not in parallel?

I would say the opposite, pastas tries parallel, unless you tell it not to. If pastastore parallel is affected by this, pastastore can turn off pastas parallel.

> Do we need if __name__ == "__main__" statements to ensure memory safety?

For numba prange we don't need this.

> How about linux/macos/windows differences for parallel?

Numba takes care of this.


The above questions are valid if we start implementing multiprocessing or multithreading within pastas, but as long as we remain in the numba parallel space, these things are taken care of by numba and we don't really need to worry. 

So I would say, go ahead on the numba front, but if we want to start implementing other forms of parallel computing, we shoud, definitely revisit these questions.