@martinvonk 2025-03-06
But can we then do something like this in this PR:
`@njit(parallel=ps.rcParams["parallel"])`
and 
```python
if ps.rcParams["parallel"]:`
    from numba import prange as range
```