@dbrakenhoff 2025-03-06
I think it would make sense to include a pastas parallel setting so it can be turned off on the module level. Unfortunately this only works on import, so you can never configure this value after importing pastas. So you would have to modify the rcparams.py file manually, which is a nice option to have, but doesn't allow interactive configuration.

The solution is to use `numba.set_num_threads(1)` if you want to toggle between serial/parallel execution using numba after importing pastas. 

Also `numba.prange` defaults to `range` when parallel is off.